requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-allauth==0.60.1
django-environ==0.11.2
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
gunicorn==21.2.0
whitenoise==6.6.0
